Move code to decode_db.py file. Remove test file.
